requirements.txt
Jinja2==2.11.2
PyYAML==5.3.1
pylint==1.8.1
ipaddress==1.0.23
